Limiter: Add lookahead implementation
Add the possibility to use lookahead when applying the limiter.
The algorithm is based on the paper referenced in [1] and is based
on using a max filter to apply the correct amount of limiting
without causing distortion.

To avoid distorted waveforms, a lookahead value of
N = Fs/(2*F) samples is needed, where Fs is the sampling rate and
F is the lowest frequency to pass the limiter without noticeable
distortion.

Only the first part of the paper is implemented, therefore no
averaging is performed.

[1] = https://users.aalto.fi/~hamalap5/dafx2002/dafx_hamalainen.pdf